Splitting main.go to multiple files to simplify file size
